+++ src/Utility/AbstractBackupUtility.php
@@ -57,10 +57,11 @@ abstract class AbstractBackupUtility
      * It provides all `getX()` methods to get properties.
      *
      * @param string $name
-     * @param array<mixed> $arguments
+     * @param array $arguments
      * @return mixed
      * @since 2.14.0
      * @throws \BadMethodCallException With a no existing property or method.
+     * @phpstan-ignore missingType.iterableValue
      */
     public function __call(string $name, array $arguments = []): mixed
     {
@@ -195,8 +196,8 @@ abstract class AbstractBackupUtility
      */
     protected function getProcess(string $command): Process
     {
-        $Process = Process::fromShellCommandline($command);
-        $Process->setTimeout($this->getTimeout() ?? Configure::readOrFail('DatabaseBackup.processTimeout'));
+        $Process = Process::fromShellCommandline(command: $command);
+        $Process->setTimeout(timeout: $this->getTimeout() ?? Configure::readOrFail('DatabaseBackup.processTimeout'));
         $Process->run();
 
         return $Process;